telegram_bot/requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
APScheduler==3.11.0
asyncpg==0.30.0
celery==5.5.3
email_validator==2.3.0
fastapi==0.116.1
httpx==0.28.1
openpyxl==3.1.3
pydantic-settings==2.10.1
PyYAML==6.0.2
redis==6.3.0
uvicorn==0.35.0
yookassa==3.7.1
